Add files via upload
Same as the "Bounce Attempt" Code except now the ball bounces correctly. The ball does go through the ceiling a little bit, but that can be fixed if we create a displacement ceiling/limit in the code. Time "t" is still used, we're going to get rid of it but the gravity bouncing concept is correct.