Half Functionality, Does Not Work But Closer
It compiles at least.
Many functions have been altered by being cut and reordered. 
Definitions of headcenter and prev_x prev_y are wrong, which is causing headvelocity_x and headvelocity_y to be 0 very often. I wanted to have the blank case use None, but Python isn't letting me say that the original headcenter is None, bc then they say that headcenter is not defined.
I think headvelocity_x is sometimes not 0, and that the value can change once before remaining static permanently. This in turn affects xvelocity and yvelocity, which I believe is causing the displacement every time the head touches the ball.

Additional problems are the weird movement paths upon contact, there are traces of the old problem of the ball moving in the opposite x direction at times, which is weird.

In addition, with 2 people the ball teleports between players, which is very strange. We will need a function to seek input from the closer player or something to differentiate the players.