Two Player Functionality Not Quite Working, but Improved
Something is affecting the ball.x_position and ball.y_position updates, and after each reset the ball mystically teleports down, likely due to some response from a huge, uncleared velocity being added onto the ball position too soon. The ball is moving in an almost straight line which is strange.

Several functions edited to fit the purpose better and to work.
Essentially rewrote the else statement for the if hit statements, allowed it to independently adjust and update velocities and ball position.